Add Resource panels
